+++ test-resources/babashka/file_location1.clj
@@ -1,0 +1,8 @@
+(ns babashka.file-location1
+  (:require [babashka.classpath :as cp]))
+
+(cp/add-classpath "test-resources")
+
+(require '[babashka.file-location2 :as f2])
+
+(f2/uh-oh)